la ricetta può esser cancellata solo dal cuoco o  un admin

+++ src/main/java/it/uniroma3/siwFood/service/RecipeService.java
@@ -3,8 +3,13 @@ package it.uniroma3.siwFood.service;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
+import it.uniroma3.siwFood.model.Cook;
+import it.uniroma3.siwFood.model.Credentials;
 import it.uniroma3.siwFood.model.Recipe;
 import it.uniroma3.siwFood.repository.RecipeRepository;
 
@@ -13,6 +18,13 @@ public class RecipeService {
 	
 	@Autowired
 	private RecipeRepository recipeRepository;
+	
+	@Autowired 
+	private CredentialsService credentialsService;
+	
+	@Autowired
+	private CookService cookService;
+	
 	
 	public List<Recipe> findAllRecipes(){
 		return (List<Recipe>) this.recipeRepository.findAll();
@@ -38,8 +50,17 @@ public class RecipeService {
 		this.recipeRepository.save(recipe);
 	}
 	
-	public void deleteRecipe(Recipe recipe) {
-		this.recipeRepository.delete(recipe);
+	public void deleteRecipe(Recipe recipe) throws AccessDeniedException{
+		
+	    UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		Credentials credentials = this.credentialsService.findCredenzialiByUsername(userDetails.getUsername());
+		Cook cook = this.cookService.findCookByCredentials(credentials.getIdCredentials());
+		
+		if(recipe.getCook().equals(cook) || credentials.getRole().equals("ADMIN")) {
+			this.recipeRepository.delete(recipe);
+		}
+		else {
+			throw new AccessDeniedException("You do not have permission to remove this recipe");
+		}
 	}
-
 }
